Initialize IDE component registration earlier
